+++ frontend/src/views/settings/Profile.vue
@@ -12,11 +12,13 @@
               class="item"
               v-model="localuser.disableQuickToggles"
               :name="$t('profileSettings.disableQuickToggles')"
+              :description="$t('profileSettings.disableQuickTogglesDescription')"
             />
             <ToggleSwitch
               class="item"
               v-model="localuser.preview.disableHideSidebar"
               :name="$t('profileSettings.disableHideSidebar')"
+              :description="$t('profileSettings.disableHideSidebarDescription')"
             />
             <ToggleSwitch
               class="item"
@@ -33,6 +35,7 @@
               class="item"
               v-model="localuser.deleteWithoutConfirming"
               :name="$t('profileSettings.deleteWithoutConfirming')"
+              :description="$t('profileSettings.deleteWithoutConfirmingDescription')"
             />
             <ToggleSwitch
               class="item"
@@ -43,44 +46,51 @@
               class="item"
               v-model="localuser.showHidden"
               :name="$t('profileSettings.showHiddenFiles')"
+              :description="$t('profileSettings.showHiddenFilesDescription')"
             />
             <ToggleSwitch
               class="item"
               v-model="localuser.quickDownload"
               :name="$t('profileSettings.showQuickDownload')"
+              :description="$t('profileSettings.showQuickDownloadDescription')"
             />
             <ToggleSwitch
               class="item"
               v-model="localuser.preview.image"
               :name="$t('profileSettings.previewImages')"
+              :description="$t('profileSettings.previewImagesDescription')"
             />
             <ToggleSwitch
               v-if="mediaEnabled"
               class="item"
               v-model="localuser.preview.video"
               :name="$t('profileSettings.previewVideos')"
+              :description="$t('profileSettings.previewVideosDescription')"
             />
             <ToggleSwitch
               v-if="mediaEnabled"
               class="item"
               v-model="localuser.preview.motionVideoPreview"
               :name="$t('profileSettings.previewMotionVideos')"
+              :description="$t('profileSettings.previewMotionVideosDescription')"
             />
             <ToggleSwitch
               class="item"
               v-model="localuser.preview.highQuality"
               :name="$t('profileSettings.highQualityPreview')"
+              :description="$t('profileSettings.highQualityPreviewDescription')"
             />
             <ToggleSwitch
-              v-if="hasOnlyOfficeEnabled"
               class="item"
               v-model="localuser.preview.office"
               :name="$t('profileSettings.previewOffice')"
+              :description="$t('profileSettings.previewOfficeDescription')"
             />
             <ToggleSwitch
               class="item"
               v-model="localuser.preview.popup"
               :name="$t('profileSettings.popupPreview')"
+              :description="$t('profileSettings.popupPreviewDescription')"
             />
           </div>
           <h3>{{ $t("profileSettings.editorViewerOptions") }}</h3>
@@ -89,11 +99,13 @@
                 class="item"
                 v-model="localuser.preview.autoplayMedia"
                 :name="$t('profileSettings.autoplayMedia')"
+                :description="$t('profileSettings.autoplayMediaDescription')"
               />
               <ToggleSwitch
                 class="item"
                 v-model="localuser.editorQuickSave"
                 :name="$t('profileSettings.editorQuickSave')"
+                :description="$t('profileSettings.editorQuickSaveDescription')"
               />
           </div>
           <h3>{{ $t("settings.searchOptions") }}</h3>
@@ -102,6 +114,7 @@
               class="item"
               v-model="localuser.disableSearchOptions"
               :name="$t('profileSettings.disableSearchOptions')"
+              :description="$t('profileSettings.disableSearchOptionsDescription')"
             />
            </div>
            <h3 v-if="user.permissions.admin">{{ $t("settings.adminOptions") }}</h3>
@@ -111,50 +124,56 @@
                 class="item"
                 v-model="localuser.disableUpdateNotifications"
                 :name="$t('profileSettings.disableUpdateNotifications')"
+                :description="$t('profileSettings.disableUpdateNotificationsDescription')"
               />
           </div>
-          <div v-if="hasOnlyOfficeEnabled">
-            <h3>{{ $t("settings.disableOfficePreview") }}</h3>
-            <p>
-              {{ $t("settings.disableOfficePreviewDescription") }}
-            </p>
+          <div>
+            <div class="centered-with-tooltip">
+              <h3>{{ $t("profileSettings.disableThumbnailPreviews") }}</h3>
+              <i class="no-select material-symbols-outlined tooltip-info-icon"
+                @mouseenter="showTooltip($event, $t('profileSettings.disableThumbnailPreviewsDescription'))" @mouseleave="hideTooltip">
+                help
+              </i>
+            </div>
             <div class="form-group">
               <input
                 class="input input--block form-form flat-right"
-                :class="{ 'invalid-form': !formValidation() }"
+                :class="{ 'invalid-form': !validateExtensions(localuser.disablePreviewExt) }"
                 type="text"
                 placeholder="enter file extensions"
-                id="onlyofficeExt"
-                v-model="formOnlyOfficeExt"
+                id="disablePreviews"
+                v-model="localuser.disablePreviewExt"
               />
               <button
                 type="button"
                 class="button form-button"
-                @click="submitOnlyOfficeChange"
+                @click="updateSettings"
               >
                 {{ $t("buttons.save") }}
               </button>
             </div>
           </div>
-
-          <div v-if="muPdfAvailable">
-            <h3>{{ $t("settings.disableOfficePreviews") }}</h3>
-            <p>
-              {{ $t("settings.disableOfficePreviewsDescription") }}
-            </p>
+          <div>
+            <div class="centered-with-tooltip">
+              <h3>{{ $t("profileSettings.disableViewingFiles") }}</h3>
+              <i class="no-select material-symbols-outlined tooltip-info-icon"
+                @mouseenter="showTooltip($event, $t('profileSettings.disableViewingFilesDescription'))" @mouseleave="hideTooltip">
+                help
+              </i>
+            </div>
             <div class="form-group">
               <input
                 class="input input--block form-form flat-right"
-                :class="{ 'invalid-form': !formValidationOfficePreviews() }"
+                :class="{ 'invalid-form': !validateExtensions(localuser.disabledViewingExt) }"
                 type="text"
                 placeholder="enter file extensions"
-                id="officePreviewExt"
-                v-model="formOfficePreviewExt"
+                id="disableViewing"
+                v-model="localuser.disabledViewingExt"
               />
               <button
                 type="button"
                 class="button form-button"
-                @click="submitOfficePreviewsChange"
+                @click="updateSettings"
               >
                 {{ $t("buttons.save") }}
               </button>
@@ -173,6 +192,9 @@
             @update:locale="updateLocale"
           ></Languages>
         </div>
+        <div class="card-action">
+          <button class="button button--flat" @click="updateSettings">{{ $t("buttons.save") }}</button>
+        </div>
       </form>
     </div>
   </div>
@@ -180,7 +202,7 @@
 
 <script>
 import { notify } from "@/notify";
-import { onlyOfficeUrl, mediaAvailable, muPdfAvailable } from "@/utils/constants.js";
+import { mediaAvailable, muPdfAvailable } from "@/utils/constants.js";
 import { state, mutations } from "@/store";
 import { usersApi } from "@/api";
 import Languages from "@/components/settings/Languages.vue";
@@ -198,8 +220,6 @@ export default {
     return {
       localuser: { preview: {}, permissions: {} }, // Initialize localuser with empty objects to avoid undefined errors
       initialized: false,
-      formOnlyOfficeExt: "", // holds temporary input before saving
-      formOfficePreviewExt: "", // holds temporary input before saving
       colorChoices: [
         { label: this.$t("colors.blue"), value: "var(--blue)" },
         { label: this.$t("colors.red"), value: "var(--red)" },
@@ -210,26 +230,12 @@ export default {
       ],
     };
   },
-  watch: {
-    localuser: {
-      handler: function () {
-        if (this.initialized) {
-          this.updateSettings(); // Ensure updateSettings() is called when localuser changes
-        }
-        this.initialized = true;
-      },
-      deep: true, // Watch nested properties of localuser
-    },
-  },
   computed: {
     user() {
       return state.user;
     },
     muPdfAvailable() {
       return muPdfAvailable;
-    },
-    hasOnlyOfficeEnabled() {
-      return onlyOfficeUrl != "";
     },
     mediaEnabled() {
       return mediaAvailable;
@@ -242,38 +248,25 @@ export default {
     },
   },
   mounted() {
-    this.localuser = { ...state.user };
-    this.formOnlyOfficeExt = this.localuser.disableOnlyOfficeExt;
-    this.formOfficePreviewExt = this.localuser.disableOfficePreviewExt;
+    this.localuser = JSON.parse(JSON.stringify(state.user));
   },
   methods: {
-    formValidation() {
-      if (this.formOnlyOfficeExt == "") {
+    showTooltip(event, text) {
+      mutations.showTooltip({
+        content: text,
+        x: event.clientX,
+        y: event.clientY,
+      });
+    },
+    hideTooltip() {
+      mutations.hideTooltip();
+    },
+    validateExtensions(value) {
+      if (value === "") {
         return true;
       }
-      let regex = /^\.\w+(?: \.\w+)*$/;
-      return regex.test(this.formOnlyOfficeExt);
-    },
-    submitOnlyOfficeChange() {
-      if (!this.formValidation()) {
-        notify.showError("Invalid input, does not match requirement.");
-        return;
-      }
-      this.localuser.disableOnlyOfficeExt = this.formOnlyOfficeExt;
-    },
-    formValidationOfficePreviews() {
-      if (this.formOfficePreviewExt == "") {
-        return true;
-      }
-      let regex = /^\.\w+(?: \.\w+)*$/;
-      return regex.test(this.formOfficePreviewExt);
-    },
-    submitOfficePreviewsChange() {
-      if (!this.formValidationOfficePreviews()) {
-        notify.showError("Invalid input, does not match requirement.");
-        return;
-      }
-      this.localuser.disableOfficePreviewExt = this.formOfficePreviewExt;
+      const regex = /^\.\w+(?: \.\w+)*$/;
+      return regex.test(value);
     },
     setColor(string) {
       this.localuser.themeColor = string;
@@ -297,8 +290,8 @@ export default {
           "dateFormat",
           "themeColor",
           "quickDownload",
-          "disableOnlyOfficeExt",
-          "disableOfficePreviewExt",
+          "disablePreviewExt",
+          "disabledViewingExt",
           "deleteWithoutConfirming",
           "preview",
           "disableQuickToggles",
@@ -322,7 +315,14 @@ export default {
 .card-content h3 {
   text-align: center;
 }
-#officePreviewExt {
+#disablePreviews,
+#disableViewing {
   width: 80%;
+}
+
+.centered-with-tooltip {
+  display: flex;
+  justify-content: center;
+  align-items: center;
 }
 </style>